added format support
ollama 0.1.9 added format support for now it's just for json

what i have changed.
request.rs
added the enum FormatEnum
added it as an option to GenerationRequest
added it in the new method for GenerationRequest

Also added a test for generating Json

updated README and examples to reflect these changes

The reason why i made it an enum instead of an bool is because they might add more format options and adding it as an enum should be easier in that case.